+++ radix-engine-common/src/crypto/bls12381/private_key.rs
@@ -39,8 +39,6 @@ impl BlsPrivateKey {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::validation::verify_bls;
-    use radix_engine_interface::crypto::hash;
     use sbor::rust::str::FromStr;
 
     #[test]
