+++ radix-engine-common/src/crypto/ed25519/mod.rs
@@ -1,5 +1,7 @@
 mod private_key;
+mod public_key;
 mod signature;
 
 pub use private_key::*;
+pub use public_key::*;
 pub use signature::*;
