+++ radix-engine-common/src/crypto/mod.rs
@@ -1,17 +1,25 @@
 mod blake2b;
+mod bls12381;
+mod ed25519;
 mod hash;
 mod hash_accumulator;
+mod private_key;
 mod public_key;
-mod public_key_bls;
-mod public_key_ed25519;
 mod public_key_hash;
-mod public_key_secp256k1;
+mod secp256k1;
+mod signature;
+mod signature_validator;
+mod signer;
 
 pub use self::blake2b::*;
+pub use self::bls12381::*;
+pub use self::ed25519::*;
 pub use self::hash::*;
 pub use self::hash_accumulator::*;
+pub use self::private_key::*;
 pub use self::public_key::*;
-pub use self::public_key_bls::*;
-pub use self::public_key_ed25519::*;
 pub use self::public_key_hash::*;
-pub use self::public_key_secp256k1::*;
+pub use self::secp256k1::*;
+pub use self::signature::*;
+pub use self::signature_validator::*;
+pub use self::signer::*;
