+++ radix-engine-common/src/crypto/secp256k1/private_key.rs
@@ -56,8 +56,6 @@ impl Secp256k1PrivateKey {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::validation::verify_secp256k1;
-    use radix_engine_interface::crypto::hash;
     use sbor::rust::str::FromStr;
 
     #[test]
