+++ radix-engine-common/src/crypto/secp256k1/public_key.rs
@@ -1,4 +1,3 @@
-use super::*;
 use crate::internal_prelude::*;
 #[cfg(feature = "radix_engine_fuzzing")]
 use arbitrary::Arbitrary;
