+++ transaction/Cargo.toml
@@ -10,9 +10,6 @@ radix-engine-interface = { path = "../radix-engine-interface", default-features 
 radix-engine-common = { path = "../radix-engine-common", default-features = false }
 
 hex = { version = "0.4.3", default-features = false }
-ed25519-dalek = { version = "1.0.1", default-features = false, features = ["u64_backend"]}
-secp256k1 = { version = "0.24.0", default-features = false, features = ["global-context", "recovery"]}
-blst = { version = "0.3.11", default-features = false }
 serde = { version = "1.0.144", default-features = false, optional = true }
 lazy_static = "1.4.0"
 strum = { version = "0.24", default-features = false, features = ["derive"] }
@@ -25,8 +22,8 @@ scrypto-derive = { path = "../scrypto-derive" }
 [features]
 # You should enable either `std` or `alloc`
 default = ["std"]
-std = ["sbor/std", "utils/std", "radix-engine-interface/std", "radix-engine-common/std", "hex/std", "ed25519-dalek/std", "secp256k1/std"]
-alloc = ["sbor/alloc", "utils/alloc", "radix-engine-interface/alloc", "radix-engine-common/alloc", "hex/alloc", "ed25519-dalek/alloc", "secp256k1/alloc", "lazy_static/spin_no_std"]
+std = ["sbor/std", "utils/std", "radix-engine-interface/std", "radix-engine-common/std", "hex/std"]
+alloc = ["sbor/alloc", "utils/alloc", "radix-engine-interface/alloc", "radix-engine-common/alloc", "hex/alloc", "lazy_static/spin_no_std"]
 serde = ["serde/derive"]
 
 dump_manifest_to_file = []
