+++ transaction/src/builder/transaction_builder.rs
@@ -1,5 +1,5 @@
+use crate::internal_prelude::*;
 use crate::model::*;
-use crate::signing::Signer;
 
 pub struct TransactionBuilder {
     manifest: Option<TransactionManifestV1>,
@@ -122,7 +122,6 @@ mod tests {
 
     use super::*;
     use crate::builder::*;
-    use crate::signing::secp256k1::Secp256k1PrivateKey;
 
     #[test]
     fn notary_as_signatory() {
