+++ transaction/src/lib.rs
@@ -3,7 +3,6 @@ pub mod data;
 pub mod errors;
 pub mod manifest;
 pub mod model;
-pub mod signing;
 pub mod validation;
 
 /// Each module should have its own prelude, which:
@@ -19,10 +18,6 @@ pub mod prelude {
     // Exports from this crate
     pub use crate::builder::*;
     pub use crate::model::*;
-    pub use crate::signing::bls::*;
-    pub use crate::signing::ed25519::*;
-    pub use crate::signing::secp256k1::*;
-    pub use crate::signing::{PrivateKey, Signer};
 }
 
 // Extra things which this crate wants which upstream crates likely don't
