+++ transaction/src/validation/mod.rs
@@ -1,9 +1,7 @@
 mod id_allocator;
 mod id_validator;
-mod signature_validator;
 mod transaction_validator;
 
 pub use id_allocator::*;
 pub use id_validator::*;
-pub use signature_validator::*;
 pub use transaction_validator::*;
